fix: make LZW more robust
While the algorithm to getNextCode was rewritten inspired by a java code
(and looks much better and faster) the main difference with the previous
implementation is this return 257 in case of OutOfBounds.

https://github.com/sugark/Tiffus/blob/master/src/org/eclipse/swt/internal/image/TIFFLZWDecoder.java#L238-L241

This change will prevent issues with probably corrupted files.
Some other implementations also allow to check for this:
https://gitlab.com/libtiff/libtiff/blob/master/libtiff/tif_lzw.c#L170-185